--- app/src/main/java/com/djangoroid/android/hackathon/UserViewModel.kt
@@ -1,4 +1,0 @@
-package com.djangoroid.android.hackathon
-
-class UserViewModel {
-}
