+++ app/src/main/java/com/djangoroid/android/hackathon/ui/user/UserViewModel.kt
@@ -1,0 +1,4 @@
+package com.djangoroid.android.hackathon.ui.user
+
+class UserViewModel {
+}
