+++ app/src/main/res/layout/fragment_login.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/fragment_vertical_margin"
     android:background="@color/white"
-    tools:context=".LoginFragment">
+    tools:context=".ui.user.LoginFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
