+++ app/src/main/res/layout/fragment_signup.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/fragment_vertical_margin"
     android:background="@color/white"
-    tools:context=".SignupFragment">
+    tools:context=".ui.user.SignupFragment">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
